+++ src/builder.rs
@@ -672,9 +672,9 @@ pub mod testing {
             for note in notes.iter() {
                 let leaf = MerkleHashOrchard::from_cmx(&note.commitment().into());
                 tree.append(&leaf);
-                let (position, leaf) = tree.witness().unwrap();
+                let (position, leaf) = tree.witness().expect("tree is not empty");
 
-                let path = MerklePath::from((position, tree.authentication_path(position, &leaf).unwrap()));
+                let path = MerklePath::from((position, tree.authentication_path(position, &leaf).expect("we just witnessed the path")));
                 notes_and_auth_paths.push((*note, path));
             }